redis_implementationV1/requirements.txt
django
django-redis
django
django-cors-headers
djangorestframework
django-filter
DateTime
requests